+++ tests/test_strings.py
@@ -57,6 +57,16 @@ def test_chars(case, expected):
     assert _.chars(case) == expected
 
 
+@parametrize('case,expected', [
+    (('foobar', 'o'), 2),
+    (('foobar', 'oo'), 1),
+    (('foobar', 'ooo'), 0),
+])
+def test_count_substr(case, expected):
+    assert _.count_substr(*case) == expected
+
+
+@parametrize('case,expected', [
     ('\xC0\xC1\xC2\xC3\xC4\xC5\xC6\xC7\xC8\xC9\xCA\xCB\xCC\xCD\xCE\xCF'
      '\xD0\xD1\xD2\xD3\xD4\xD5\xD6\xD7\xD8\xD9\xDA\xDB\xDC\xDD\xDE\xDF'
      '\xE0\xE1\xE2\xE3\xE4\xE5\xE6\xE7\xE8\xE9\xEA\xEB\xEC\xED\xEE\xEF'