+++ pydash/__init__.py
@@ -325,6 +325,7 @@ from .strings import (
     trim_right,
     trunc,
     unescape,
+    unquote,
     url,
     words,
 )