+++ CHANGES.rst
@@ -6,6 +6,7 @@ v3.0.0 (xxxx-xx-xx)
 -------------------
 
 - Add ``chop``.
+- Add ``clean``.
 - Add ``decapitalize``.
 - Make ``camel_case`` work with empty string.