+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 v3.0.0 (xxxx-xx-xx)
 -------------------
 
+- Add ``chars``.
 - Add ``chop``.
 - Add ``clean``.
 - Add ``decapitalize``.