+++ CHANGES.rst
@@ -14,6 +14,7 @@ v3.0.0 (xxxx-xx-xx)
 - Add ``insert_substr``.
 - Add ``is_blank``.
 - Add ``lines``.
+- Add ``predecessor``, ``pred``.
 - Add ``replace``.
 - Add ``swap_case``.
 - Make ``camel_case`` work with empty string.