+++ CHANGES.rst
@@ -16,6 +16,7 @@ v3.0.0 (xxxx-xx-xx)
 - Add ``lines``.
 - Add ``predecessor``, ``pred``.
 - Add ``replace``.
+- Add ``successor``, ``succ``.
 - Add ``swap_case``.
 - Make ``camel_case`` work with empty string.