+++ CHANGES.rst
@@ -19,6 +19,7 @@ v3.0.0 (xxxx-xx-xx)
 - Add ``successor``.
 - Add ``swap_case``.
 - Add ``title_case``.
+- Add ``to_boolean``.
 - Add ``to_number``.
 - Make ``camel_case`` work with empty string.
 - Make ``splice`` work with strings.